reduction on server bundle file using webpack-node-externals and merging webpack config files in server and client using webpack-merge
