+++ webpack.base.js
@@ -1,0 +1,19 @@
+module.exports = {
+    // tell webpack to run babel in every file it runs through
+    module: {
+        rules: [
+            {
+                test: /\.js?$/,
+                loader: 'babel-loader',
+                exclude: /node_modules/,
+                options: {
+                    presets: [
+                        'react',
+                        'stage-0',
+                        ['env', {targets: {browsers: ['last 2 versions']}}]
+                    ]
+                }
+            }
+        ]
+    }
+}
