+++ webpack.client.js
@@ -1,6 +1,8 @@
 const path = require('path')
+const merge = require('webpack-merge')
+const baseConfig = require('./webpack.base')
 
-module.exports = {
+const config = {
     entry: './src/client/index.js',
     // tell webpack where to put file
     output: {
@@ -8,23 +10,6 @@ module.exports = {
         path: path.resolve(__dirname, 'public')
     },
 
-    // tell webpack to run babel in every file it runs through
-
-    module: {
-        rules: [
-            {
-                test: /\.js?$/,
-                loader: 'babel-loader',
-                exclude: /node_modules/,
-                options: {
-                    presets: [
-                        'react',
-                        'stage-0',
-                        ['env', {targets: {browsers: ['last 2 versions']}}]
-                    ]
-                }
-            }
-        ]
-    }
-
 }
+
+module.exports = merge(baseConfig, config)
