+++ webpack.server.js
@@ -1,6 +1,9 @@
 const path = require('path')
+const merge = require('webpack-merge')
+const baseConfig = require('./webpack.base')
+const webpackNodeExternals = require('webpack-node-externals')
 
-module.exports = {
+const config = {
     target: 'node',
     entry: './src/index.js',
     // tell webpack where to put file
@@ -8,24 +11,7 @@ module.exports = {
         filename: 'bundle.js',
         path: path.resolve(__dirname, 'build')
     },
-
-    // tell webpack to run babel in every file it runs through
-
-    module: {
-        rules: [
-            {
-                test: /\.js?$/,
-                loader: 'babel-loader',
-                exclude: /node_modules/,
-                options: {
-                    presets: [
-                        'react',
-                        'stage-0',
-                        ['env', {targets: {browsers: ['last 2 versions']}}]
-                    ]
-                }
-            }
-        ]
-    }
-
+    externals: [webpackNodeExternals()]
 }
+
+module.exports = merge(baseConfig, config)
